Specify files to publish to npm
Currently the `example` and `test` folders and various dotfiles such as `.travis.yml` are being published to npm. Considering that they make up about half of the file size of this package, it would be worthwhile to exclude them.

This PR excludes them using the using the [`files`](https://docs.npmjs.com/files/package.json#files) whitelist in the `package.json`.